GoogleSearchWithLLM/requirements.txt
accelerate==0.29.2
albumentations==1.4.3
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
huggingface-hub==0.22.2
idna==3.4
imageio==2.34.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==23.3.1
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.7.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
watchdog==4.0.0
wheel==0.41.2
zipp==3.18.1
accelerate==0.29.2
albumentations==1.4.3
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
huggingface-hub==0.22.2
idna==3.4
imageio==2.34.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==23.3.1
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.7.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
watchdog==4.0.0
wheel==0.41.2
zipp==3.18.1